#1737: First pass on refactoring cli.js
