+++ lib/cli.js
@@ -2,104 +2,10 @@
 
 // Modules
 const _ = require('lodash');
+const formatters = require('./formatters');
 const fs = require('fs');
 const os = require('os');
 const path = require('path');
-const util = require('util');
-
-/*
- * Helper to sort options
- */
-const sortOptions = options => _(options)
-  .keys()
-  .sortBy()
-  .map(key => [key, options[key]])
-  .fromPairs()
-  .value();
-
-/*
- * Helper to get interactive options
- */
-const getInteractive = (options, argv) => _(options)
-  .map((option, name) => _.merge({}, {name}, {option}))
-  .filter(option => !_.isEmpty(_.get(option, 'option.interactive', {})))
-  .map(option => _.merge({}, {name: option.name, weight: 0}, option.option.interactive))
-  .map(option => {
-    if (_.isNil(argv[option.name]) || argv[option.name] === false) return option;
-    else {
-      return _.merge({}, option, {when: answers => {
-        answers[option.name] = argv[option.name];
-        return false;
-      }});
-    }
-  })
-  .value();
-
-/*
- * Helper to check for updates
- */
-const checkForUpdates = lando => {
-  // Check for updates
-  return lando.Promise.try(() => {
-    if (lando.updates.fetch(lando.cache.get('updates'))) {
-      lando.log.verbose('Checking for updates...');
-      return lando.updates.refresh(lando.config.version)
-      .then(latest => lando.cache.set('updates', latest, {persist: true}));
-    }
-  })
-  // Update warning
-  .then(() => {
-    if (lando.updates.updateAvailable(lando.config.version, lando.cache.get('updates').version)) {
-      console.error(lando.cli.makeArt('updateAvailable', lando.cache.get('updates').url));
-    }
-  });
-};
-
-/*
- * Helper to clear tasks cache
- */
-const clearTaskCaches = () => {
-  if (fs.existsSync(process.landoTaskCacheFile)) fs.unlinkSync(process.landoTaskCacheFile);
-  if (fs.existsSync(process.landoAppTaskCacheFile)) fs.unlinkSync(process.landoAppTaskCacheFile);
-};
-
-/*
- * Helper to prompt the user if needed
- */
-const handleInteractive = (inquiry, argv, command, lando) => lando.Promise.try(() => {
-  if (_.isEmpty(inquiry)) return {};
-  else {
-    const inquirer = require('inquirer');
-    inquirer.registerPrompt('autocomplete', require('inquirer-autocomplete-prompt'));
-    // Try to rebuild the inquiry if this is app level bootstrap and we have an app
-    if (!_.isEmpty(argv._app) && lando._bootstrap === 'app') {
-      // NOTE: We need to clone deep here otherwise any apps with interactive options get 2x all their events
-      // NOTE: Not exactly clear on why app here gets conflated with the app returned from lando.getApp
-      const app = _.cloneDeep(lando.getApp(argv._app.root));
-      return app.init().then(() => {
-        inquiry = getInteractive(_.find(app.tasks.concat(lando.tasks), {command: command}).options, argv);
-        return inquirer.prompt(_.sortBy(inquiry, 'weight'));
-      });
-    // Otherwise just run
-    } else {
-      inquiry = getInteractive(_.find(lando.tasks, {command: command}).options, argv);
-      return inquirer.prompt(_.sortBy(inquiry, 'weight'));
-    }
-  }
-});
-
-/*
- * Toggle the secret toggle
- * and return its current state
- */
-const toggleSecretToggle = toggleFile => {
-  if (fs.existsSync(toggleFile)) {
-    fs.unlinkSync(toggleFile);
-  } else {
-    fs.closeSync(fs.openSync(toggleFile, 'w'));
-  }
-  return fs.existsSync(toggleFile);
-};
 
 /*
  * Construct the CLI
@@ -140,12 +46,31 @@ module.exports = class Cli {
   };
 
   /*
+   * Helper to check for updates
+   */
+  checkForUpdates(lando) {
+    // Check for updates
+    return lando.Promise.try(() => {
+      if (lando.updates.fetch(lando.cache.get('updates'))) {
+        lando.log.verbose('Checking for updates...');
+        return lando.updates.refresh(lando.config.version)
+        .then(latest => lando.cache.set('updates', latest, {persist: true}));
+      }
+    })
+    // Update warning
+    .then(() => {
+      if (lando.updates.updateAvailable(lando.config.version, lando.cache.get('updates').version)) {
+        console.error(lando.cli.makeArt('updateAvailable', lando.cache.get('updates').url));
+      }
+    });
+  };
+
+  /*
    * Toggle the secret toggle
    */
   clearTaskCaches() {
-    clearTaskCaches();
-    console.log('Lando has cleared the tasks cache!');
-    process.exit(0);
+    if (fs.existsSync(process.landoTaskCacheFile)) fs.unlinkSync(process.landoTaskCacheFile);
+    if (fs.existsSync(process.landoAppTaskCacheFile)) fs.unlinkSync(process.landoAppTaskCacheFile);
   };
 
   /*
@@ -207,41 +132,15 @@ module.exports = class Cli {
   /*
    * Format data
    */
-  formatData(input, {path = '', format = 'default', filter = {}} = {}) {
-    const data = path && _.has(input, path) ? _.get(input, path) : input;
-    switch (format) {
-        case 'json':
-          return JSON.stringify(data);
-          break;
-        // @TODO: Add CSV.
-        default:
-          return util.inspect(data, {
-            colors: process.stdout.isTTY,
-            depth: 10,
-            compact: false,
-          });
-      }
+  formatData(data, {path = '', format = 'default', filter = {}} = {}) {
+    return formatters.formatData(data, {path, format, filter});
   };
 
   /*
    * FormatOptios
    */
   formatOptions(omit = []) {
-    return _.omit({
-      format: {
-        describe: 'Output in given format: json',
-        string: true,
-      },
-      path: {
-        describe: 'Only return the value at the given path',
-        default: null,
-        string: true,
-      },
-      filter: {
-        describe: 'Filter by "key=value"',
-        array: true,
-      },
-    }, omit);
+    return formatters.formatOptions(omit);
   };
 
   /**
@@ -277,6 +176,47 @@ module.exports = class Cli {
     // Report error if user has error reporting on
     .then(() => handler.handle(error, lando.cache.get('report_errors')).then(code => process.exit(code)));
   };
+
+  /*
+   * Init
+   */
+  init(yargs, tasks, config) {
+    // Start up the CLI
+    const cmd = !_.has(process, 'pkg') ? '$0' : path.basename(_.get(process, 'execPath', 'lando'));
+    yargs.usage(`Usage: ${cmd} <command> [args] [options]`)
+      .demandCommand(1, 'You need at least one command before moving on')
+      .example('lando start', 'Run lando start')
+      .example('lando rebuild --help', 'Get help about using the lando rebuild command')
+      .example('lando destroy -y -vvv', 'Run lando destroy non-interactively and with maximum verbosity')
+      .example('lando --clear', 'Clear the lando tasks cache')
+      .middleware([(argv => {
+        argv._app = config;
+      })])
+      .recommendCommands()
+      .wrap(yargs.terminalWidth() * 0.70)
+      .option('clear', {describe: 'Clears the lando tasks cache', global: true, type: 'boolean'})
+      .help(false)
+      .option('lando', {hidden: true, type: 'boolean'})
+      .option('help', {describe: 'Shows lando or delegated command help if applicable', type: 'boolean'})
+      .option('verbose', {alias: 'v', describe: 'Runs with extra verbosity', type: 'count'})
+      .version(false);
+
+    // Loop through the tasks and add them to the CLI
+    _.forEach(_.sortBy(tasks, 'command'), task => {
+      if (_.has(task, 'handler')) yargs.command(task);
+      else yargs.command(this.parseToYargs(task));
+    });
+
+    // Show help unless this is a delegation command
+    const current = _.find(tasks, {command: yargs.argv._[0]});
+    if ((yargs.argv.help || yargs.argv.lando) && _.get(current, 'delegate', false) === false) {
+      yargs.showHelp('log');
+      process.exit(0);
+    }
+
+    // YARGZ MATEY
+    yargs.argv;
+  }
 
 
   /**
@@ -340,7 +280,7 @@ module.exports = class Cli {
   parseToYargs({command, describe, options = {}, run = {}, level = 'app'}) {
     const handler = argv => {
       // Immediately build some arg data set opts and interactive options
-      const data = {options: argv, inquiry: getInteractive(options, argv)};
+      const data = {options: argv, inquiry: formatters.getInteractive(options, argv)};
       // Summon lando
       const Lando = require('./../lib/lando');
       const lando = new Lando(this.defaultConfig());
@@ -351,7 +291,7 @@ module.exports = class Cli {
       });
 
       // Check for updates and get things started
-      return checkForUpdates(lando).then(() => lando.bootstrap(level).then(lando => {
+      return this.checkForUpdates(lando).then(() => lando.bootstrap(level).then(lando => {
         /**
          * Event that allows altering of argv or inquirer before interactive prompts
          * are run
@@ -367,7 +307,7 @@ module.exports = class Cli {
 
         // Attempt to filter out questions that have already been answered
         // Prompt the use for feedback if needed and sort by weight
-        .then(() => handleInteractive(data.inquiry, data.options, command, lando))
+        .then(() => formatters.handleInteractive(data.inquiry, data.options, command, lando))
 
         /**
          * Event that allows final altering of answers before the task runs
@@ -394,7 +334,7 @@ module.exports = class Cli {
     };
 
     // Return our yarg command
-    return {command, describe, builder: sortOptions(options), handler};
+    return {command, describe, builder: formatters.sortOptions(options), handler};
   };
 
   /*
@@ -406,58 +346,40 @@ module.exports = class Cli {
     const yargs = require('yargs');
 
     // Clear any task caches we have
-    if (yargs.argv.clear) this.clearTaskCaches();
-    // Toggle secret toggle
-    if (yargs.argv['secret-toggle']) this.secretToggle();
-
-    // Start up the CLI
-    const cmd = !_.has(process, 'pkg') ? '$0' : path.basename(_.get(process, 'execPath', 'lando'));
-    yargs.usage(`Usage: ${cmd} <command> [args] [options]`)
-      .demandCommand(1, 'You need at least one command before moving on')
-      .example('lando start', 'Run lando start')
-      .example('lando rebuild --help', 'Get help about using the lando rebuild command')
-      .example('lando destroy -y -vvv', 'Run lando destroy non-interactively and with maximum verbosity')
-      .example('lando --clear', 'Clear the lando tasks cache')
-      .middleware([(argv => {
-        argv._app = config;
-      })])
-      .recommendCommands()
-      .wrap(yargs.terminalWidth() * 0.70)
-      .option('clear', {describe: 'Clears the lando tasks cache', global: true, type: 'boolean'})
-      .help(false)
-      .option('lando', {hidden: true, type: 'boolean'})
-      .option('help', {describe: 'Shows lando or delegated command help if applicable', type: 'boolean'})
-      .option('verbose', {alias: 'v', describe: 'Runs with extra verbosity', type: 'count'})
-      .version(false);
-
-    // Loop through the tasks and add them to the CLI
-    _.forEach(_.sortBy(tasks, 'command'), task => {
-      if (_.has(task, 'handler')) yargs.command(task);
-      else yargs.command(this.parseToYargs(task));
-    });
-
-    // Show help unless this is a delegation command
-    const current = _.find(tasks, {command: yargs.argv._[0]});
-    if ((yargs.argv.help || yargs.argv.lando) && _.get(current, 'delegate', false) === false) {
-      yargs.showHelp('log');
+    if (yargs.argv.clear) {
+      this.clearTaskCaches();
+      console.log('Lando has cleared the tasks cache!');
       process.exit(0);
     }
 
-    // YARGZ MATEY
-    yargs.argv;
+    // Toggle secret toggle
+    if (yargs.argv['secret-toggle']) {
+      // Deny if we are packaged
+      if (this.defaultConfig().packaged) {
+        console.error(this.makeArt('secretToggleDenied'));
+        process.exit(1);
+      }
+
+      // Toggle
+      this.secretToggle(this.defaultConfig().secretToggle);
+      this.clearTaskCaches();
+      console.log(this.makeArt('secretToggle', fs.existsSync(this.defaultConfig().secretToggle)));
+      process.exit(0);
+    }
+
+    // Initialize
+    this.init(yargs, tasks, config);
   };
 
   /*
    * Toggle the secret toggle
    */
-  secretToggle() {
-    if (this.defaultConfig().packaged) {
-      console.error(this.makeArt('secretToggleDenied'));
-      process.exit(1);
+  secretToggle(toggleFile) {
+    if (fs.existsSync(toggleFile)) {
+      fs.unlinkSync(toggleFile);
     } else {
-      console.log(this.makeArt('secretToggle', toggleSecretToggle(this.defaultConfig().secretToggle)));
-      clearTaskCaches();
-      process.exit(0);
+      fs.closeSync(fs.openSync(toggleFile, 'w'));
     }
+    return fs.existsSync(toggleFile);
   };
 };
